requirements.txt
pytorch
cytoolz
gensim
pandas
scikit-learn
tensorboardX

